% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg.R
\name{coef.ridgereg}
\alias{coef.ridgereg}
\title{This contains the regression coefficents of ridgereg function.}
\usage{
\method{coef}{ridgereg}(object, ...)
}
\arguments{
\item{object}{An object of ridgereg class}

\item{...}{Further arguments passed to or from other methods}
}
\description{
This contains the regression coefficents of ridgereg function.
}
