% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg.R
\name{pred.ridgereg}
\alias{pred.ridgereg}
\title{This contains the fitted values of ridgereg function.}
\usage{
pred.ridgereg(p, ...)
}
\arguments{
\item{p}{An object of ridgereg class}

\item{...}{Further arguments passed to or from other methods}
}
\description{
This contains the fitted values of ridgereg function.
}
