% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg.R
\name{predict.ridgereg}
\alias{predict.ridgereg}
\title{This contains the fitted values of ridgereg function.}
\usage{
\method{predict}{ridgereg}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{An object of ridgereg class}

\item{newdata}{this is newdata}

\item{...}{Further arguments passed to or from other methods}
}
\description{
This contains the fitted values of ridgereg function.
}
