% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg.R
\name{print.ridgereg}
\alias{print.ridgereg}
\title{This contains the print methods for ridgereg function.}
\usage{
\method{print}{ridgereg}(x, ...)
}
\arguments{
\item{x}{An object of ridgereg class}

\item{...}{Further arguments passed to or from other methods}
}
\description{
This contains the print methods for ridgereg function.
}
