% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg.R
\docType{class}
\name{ridgereg-class}
\alias{ridgereg-class}
\alias{ridgereg}
\title{ridgereg}
\arguments{
\item{name}{description formula formula.}

\item{data}{data.frame.}

\item{lambda}{numeric.}
}
\value{
a ridgreg object
}
\description{
ridgereg A ridge regression function
}
\details{
Ridge Regression
}
\section{Fields}{

\describe{
\item{\code{x_matrix}}{matrix.}

\item{\code{normalized_x_matrix}}{matrix.}

\item{\code{y_matrix}}{matrix.}

\item{\code{beta_ridge}}{matrix.}

\item{\code{prediction}}{matrix.}

\item{\code{data_name}}{character.}
}}


