% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg.R
\name{ridgereg}
\alias{ridgereg}
\title{Perform ridge regression using ordinary least squares as well as QR decomposition}
\usage{
ridgereg(formula, data, lambda = 0)
}
\arguments{
\item{formula}{an object, that is symbolic description of the model to be fitted.}

\item{data}{a data frame.}

\item{lambda}{the lambda of the data set(hyperparameter). Default is 0.}
}
\value{
An object of class ridgereg
}
\description{
Perform ridge regression using ordinary least squares as well as QR decomposition
}
