% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise_airport_delay.R
\name{visualize_airport_delays}
\alias{visualize_airport_delays}
\title{Visualize Airport Delays}
\usage{
visualize_airport_delays()
}
\value{
The function returns a graph.
}
\description{
This function plots the Mean Delay Time of Flights in USA Airports.
The function does not take any parameters.
}
