% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg.R
\name{predict.ridgereg}
\alias{predict.ridgereg}
\title{Predict Method for Ridge Regression Objects}
\usage{
\method{predict}{ridgereg}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{"ridgereg"} returned by \code{\link{ridgereg}}.}

\item{newdata}{New data for prediction (data frame or matrix).}

\item{...}{Additional arguments.}
}
\value{
A numeric vector of predicted (fitted) values.
}
\description{
Returns the fitted values from a ridge regression model.
}
\examples{
model <- ridgereg(y ~ x1 + x2, data = mydata, lambda = 1)
preds <- predict(model, newdata = mydata)

}
