% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg.R
\name{predict.ridgereg}
\alias{predict.ridgereg}
\title{Predict Method for Ridge Regression Objects}
\usage{
\method{predict}{ridgereg}(x)
}
\arguments{
\item{x}{An object of class \code{"ridgereg"} returned by \code{\link{ridgereg}}.}
}
\value{
A numeric vector of predicted (fitted) values.
}
\description{
Returns the fitted values from a ridge regression model.
}
