% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.ridgereg.R
\name{predict.ridgereg}
\alias{predict.ridgereg}
\title{Predicts reponse of ridge regression}
\usage{
\method{predict}{ridgereg}(object, Xnew, ...)
}
\arguments{
\item{object}{A fitted 'ridgereg' object.}

\item{Xnew}{A matrix of new values for x at which predictions are to be made.}

\item{...}{Not used. The dot argument that is passed on to predict.}
}
\value{
A vector of fitted responses.
}
\description{
\code{predict.ridgereg} fits the response from a fitted 'ridgereg' object.
}
