% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg.R
\name{RidgeReg}
\alias{RidgeReg}
\title{Ridge regression}
\usage{
RidgeReg(X, y, lambda, stan = TRUE)
}
\arguments{
\item{X}{Covariates.}

\item{y}{Binary response.}

\item{lambda}{A sequence of tuning parameters.}

\item{stan}{Whether to standardize data.}
}
\value{
An S3 class 'ridgereg' object with the following two components:
  betas A p by L matrix containing coefficients for each of L lambda values.
  b0    A p-vector containing intercepts for each of L lambda values.
}
\description{
\code{ridge} fits ridge regression by singular value decomposition.
}
