% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_airport_delays.R
\name{visualize_airport_delays}
\alias{visualize_airport_delays}
\title{Visualize mean arrival delays for airports}
\usage{
visualize_airport_delays()
}
\description{
This function creates a scatter plot of mean arrival delays by airport longitude and latitude.
}
